requirements.txt
uvicorn
fastapi
Pillow
numpy
keras
tensorflow
python-multipart